Build/crash fix: Didn't find class "okio.Buffer"
okio is a dependency of okhttp3. Presumably, it wasn't necessary
for this dependency to be specified in other developers'
environments (perhaps due to a version difference), but explicitly
specifying this dependency in build.gradle was necessary for me.

The error message encountered was something like this:

  java.lang.ClassNotFoundException: Didn't find class "okio.Buffer"

(It was longer, but I can't be more specific because I no longer
have the log for it.)